plugintrace: ServerLogTracer function
This is a helper for simple cases where the caller just wants to see a
log of the trace events for debugging and doesn't care exactly how they
are formatted.

Callers that want more control (e.g. to produce distributed trace events,
or to produce structured logs) can still set up their own ServerTracer
to do whatever is needed.