Negotiate plugin transports, rather than assuming "unix,tcp"
Clients written in some languages will have a hard time supporting Unix
sockets because support is absent in their standard libraries, so we'll
allow clients to opt out and force the use of TCP when needed.

For compatibility with hashicorp/go-plugin clients, the server will assume
that both protocols are supported if the negotiation header isn't present.