server: Allow disabling TLS in spite of that being invalid per spec
One of the simplifications of rpcplugin as compared to HashiCorp's
go-plugin is requiring TLS in order to remove one of the decision points
from designing a plugin protocol.

However, allowing a way to intentionally violate that requirement makes
the rpcplugin server implementation compatible with more go-plugin
clients, so out of pragmatism we'll offer a way to get that behavior even
though it violates the spec.

The default behavior is still to automatically negotiate a temporary
certificate for each plugin.