server: Additional shimming for go-plugin compatibility
We previously added some light shimming to help rpcplugin clients interop
with go-plugin servers, but still had some rough edges for the opposite
situation where the client is HashiCorp go-plugin while the server is
rpcplugin.

The situation in this direction is trickier because go-plugin clients make
some special assumptions:
   1. The certificate is encoded in a nonstandard base64 encoding without
      any padding. (Parsing will fail if standard padding is present)
   2. The client expects the server to implement an additional gRPC service
      named just "plugin" which has a service function
      GRPCController.Shutdown, which is used instead of sending a signal
      because go-plugin servers ignore all incoming signals.

To implement these we use a new heuristic based on the fact that go-plugin
clients don't explicitly announce which transports they support, and
therefore a caller which appears to be trying to be a plugin client in
all other ways but lacks explicit transport announcements is probably
go-plugin, and therefore activates these shims.