gopluginshim: Actually capture the close callback
Previously this was just discarding the given close callback, making any
shutdown request cause the process to panic.

The panic output was being eaten completely in some cases anyway since
go-plugin clients tend to mostly ignore the stderr of plugin server
processes, but some of them in some cases sniff stderr for something that
"looks like" a Go panic report and display it to help a plugin developer
with debugging.

In any event, panicking when asked to close is not the correct behavior
since it's supposed to cause the context.Context in the server to be
cancelled instead.